track the number of pined safes on the accounts page
