+++ src/components/welcome/MyAccounts/useTrackedSafesCount.ts
@@ -1,22 +1,47 @@
 import { AppRoutes } from '@/config/routes'
 import { OVERVIEW_EVENTS, trackEvent } from '@/services/analytics'
 import { useRouter } from 'next/router'
-import { useEffect } from 'react'
+import { useEffect, useMemo } from 'react'
 import type { ConnectedWallet } from '@/hooks/wallets/useOnboard'
 import { type SafeItem } from './useAllSafes'
+import type { AllSafesGrouped } from './useAllSafesGrouped'
 import { type MultiChainSafeItem } from './useAllSafesGrouped'
 import { isMultiChainSafeItem } from '@/features/multichain/utils/utils'
 
 let isOwnedSafesTracked = false
+let isPinnedSafesTracked = false
 let isWatchlistTracked = false
 
 const useTrackSafesCount = (
-  ownedSafes: (MultiChainSafeItem | SafeItem)[] | undefined,
-  watchlistSafes: (MultiChainSafeItem | SafeItem)[] | undefined,
+  safes: AllSafesGrouped,
+  pinnedSafes: (MultiChainSafeItem | SafeItem)[],
   wallet: ConnectedWallet | null,
 ) => {
   const router = useRouter()
   const isLoginPage = router.pathname === AppRoutes.welcome.accounts
+
+  const ownedMultiChainSafes = useMemo(
+    () => safes.allMultiChainSafes?.filter((account) => account.safes.some(({ isWatchlist }) => !isWatchlist)),
+    [safes],
+  )
+
+  // If all safes of a multichain account are on the watchlist we put the entire account on the watchlist
+  const watchlistMultiChainSafes = useMemo(
+    () => safes.allMultiChainSafes?.filter((account) => !account.safes.some(({ isWatchlist }) => !isWatchlist)),
+    [safes],
+  )
+
+  const ownedSafes = useMemo<(MultiChainSafeItem | SafeItem)[]>(
+    () => [...(ownedMultiChainSafes ?? []), ...(safes.allSingleSafes?.filter(({ isWatchlist }) => !isWatchlist) ?? [])],
+    [safes, ownedMultiChainSafes],
+  )
+  const watchlistSafes = useMemo<(MultiChainSafeItem | SafeItem)[]>(
+    () => [
+      ...(watchlistMultiChainSafes ?? []),
+      ...(safes.allSingleSafes?.filter(({ isWatchlist }) => isWatchlist) ?? []),
+    ],
+    [safes, watchlistMultiChainSafes],
+  )
 
   // Reset tracking for new wallet
   useEffect(() => {
@@ -33,6 +58,17 @@ const useTrackSafesCount = (
       isOwnedSafesTracked = true
     }
   }, [isLoginPage, ownedSafes, wallet])
+
+  useEffect(() => {
+    const totalSafesPinned = pinnedSafes?.reduce(
+      (prev, current) => prev + (isMultiChainSafeItem(current) ? current.safes.length : 1),
+      0,
+    )
+    if (wallet && !isPinnedSafesTracked && pinnedSafes && pinnedSafes.length > 0 && isLoginPage) {
+      trackEvent({ ...OVERVIEW_EVENTS.TOTAL_SAFES_PINNED, label: totalSafesPinned })
+      isPinnedSafesTracked = true
+    }
+  }, [isLoginPage, pinnedSafes, wallet])
 
   useEffect(() => {
     const totalSafesWatched = watchlistSafes?.reduce(
